Tests pass with jest
