+++ package.json
@@ -42,6 +42,7 @@
     "typescript": "^4.3.4"
   },
   "jest": {
+    "maxWorkers": "50%",
     "testPathIgnorePatterns": [
       "<rootDir>/__tests__/utils"
     ],
